feat(map): implements map components
