+++ airsen-frontend/src/app/features/map/map-view/map-view.component.html
@@ -1,0 +1,6 @@
+<!-- Map Content -->
+<div class="map-container">
+  <div class="map-frame">
+    <div id="map"></div>
+  </div>
+</div>
