+++ airsen-frontend/src/app/features/map/map-view/map-view.component.ts
@@ -1,0 +1,32 @@
+import { Component, AfterViewInit } from '@angular/core';
+import * as L from 'leaflet';
+
+@Component({
+  standalone: false,
+  selector: 'app-map-view',
+  templateUrl: './map-view.component.html',
+  styleUrls: ['./map-view.component.scss']
+})
+export class MapViewComponent implements AfterViewInit {
+  private map: any;
+
+  ngAfterViewInit(): void {
+    this.initMap();
+  }
+
+  private initMap(): void {
+    this.map = L.map('map', {
+      // coordonnées du milieu de la france
+      center: [ 46.3622, 1.5231 ],
+      zoom: 6
+    });
+
+    const tiles = L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
+      maxZoom: 18,
+      minZoom: 3,
+      attribution: '&copy; <a href="http://www.openstreetmap.org/copyright">OpenStreetMap</a>'
+    });
+
+    tiles.addTo(this.map);
+  }
+}
