+++ airsen-frontend/src/app/features/map/map.component.html
@@ -3,21 +3,45 @@
   @if (isLoading){
   <div class="loading-container">
     <mat-spinner diameter="50"></mat-spinner>
-    <p>Loading map...</p>
+    <p>Chargement de la carte...</p>
   </div>
   } @if (!isLoading && currentUser){
   <div class="page-content">
     <div class="favorites-container">
-      <h2>Mes favoris</h2>
+      <h2>
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          stroke-width="2"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          class="lucide lucide-heart w-5 h-5 text-primary"
+          data-lov-id="src/pages/Favorites.tsx:211:18"
+          data-lov-name="Heart"
+          data-component-path="src/pages/Favorites.tsx"
+          data-component-line="211"
+          data-component-file="Favorites.tsx"
+          data-component-name="Heart"
+          data-component-content="%7B%22className%22%3A%22w-5%20h-5%20text-primary%22%7D"
+        >
+          <path
+            d="M19 14c1.49-1.46 3-3.21 3-5.5A5.5 5.5 0 0 0 16.5 3c-1.76 0-3 .5-4.5 2-1.5-1.5-2.74-2-4.5-2A5.5 5.5 0 0 0 2 8.5c0 2.3 1.5 4.05 3 5.5l7 7Z"
+          ></path>
+        </svg>
+        Mes favoris
+      </h2>
     </div>
     <!-- Map Content -->
-    <div class="map-container">
-      <div class="map-frame">
-        <div id="map"></div>
-      </div>
-      <div class="map-datas">
-        <h2>Données environnementales</h2>
-      </div>
+    <app-map-view></app-map-view>
+    <div class="map-datas">
+      <h2>Données environnementales</h2>
+      <h3>Qualité de l'air</h3>
+      <h3>Météo</h3>
+      <h3>Géo</h3>
     </div>
   </div>
   }
@@ -26,10 +50,10 @@
   @if (!isLoading && !currentUser){
   <div class="error-container">
     <mat-icon>error_outline</mat-icon>
-    <h2>Unable to load map</h2>
-    <p>Please try logging in again</p>
+    <h2>Impossible de charger la carte</h2>
+    <p>Cela peut être dû à un dysfonctionnement interne ou une erreur de connexion</p>
     <button mat-raised-button color="primary" (click)="logout()">
-      Back to Login
+      Retour au Login
     </button>
   </div>
   }
