+++ airsen-frontend/src/app/features/map/map.component.scss
@@ -1,9 +1,8 @@
 @use 'src/assets/design-system.scss' as *;
 
 .page-container {
-  background-color: $airsen-gray-50;
+  background-color: $airsen-gray-100;
   padding: 2rem;
-  min-height: 100vh;
   border-radius: 8px;
 
   // Loading State
@@ -49,46 +48,43 @@
 }
 
 .page-content{
-
+  padding: 20px;
   display: flex;
+  gap: 20px;
+  width: 100%;
 
   .favorites-container{
     color: $airsen-gray-800;
+    width:20%;
+    height: fit-content;
+    background-color: $airsen-gray-50;
+    border-radius: 8px;
+    border: 1px solid $airsen-gray-300;
+    padding: 20px;
 
     h2{
       color: $airsen-gray-900;
     }
   }
 
-  // map Content
-  .map-container {
-    width: 50%;
-    min-height: 750px;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    
-    .map-frame {
-      border: 2px solid $airsen-gray-300;
-      border-radius: 10px;
-      height: 100%;
-      width: 100%;
-      
-      #map {
-          height: 100%;
-          width: 100%;
-        }
+    .map-datas{
+      display: flex;
+      flex-direction: column;
+      gap: 1rem;
+      color: $airsen-gray-800;
+      width:30%;
+      height: fit-content;
+      background-color: $airsen-gray-50;
+      border-radius: 8px;
+      border: 1px solid $airsen-gray-300;
+      padding: 20px;
+
+      h2{
+        color: $airsen-gray-900;
       }
 
-      .map-datas{
-        margin-top: 1rem;
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-        
-        h2{
-          color: $airsen-gray-900;
-        }
+      h3{
+        text-decoration: underline;
       }
     }
   }
